requirements.txt
beautifulsoup4==4.13.4
boto3==1.38.41
botocore==1.38.41
fastapi==0.115.13
langdetect==1.0.9
llama-cloud==0.1.26
llama-cloud-services==0.6.34
llama-index==0.12.43
llama-index-agent-openai==0.4.11
llama-index-cli==0.4.3
llama-index-core==0.12.43
llama-index-embeddings-gemini==0.3.2
llama-index-embeddings-google==0.3.1
llama-index-embeddings-google-genai==0.2.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.7
llama-index-instrumentation==0.2.0
llama-index-llms-gemini==0.5.0
llama-index-llms-google-genai==0.2.1
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-postgres==0.5.3
llama-index-workflows==0.2.2
llama-parse==0.6.34
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
PyMuPDF==1.26.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
uvicorn
sqlmodel
llama-cloud-services
python-multipart
